+++ app/src/main/AndroidManifest.xml
@@ -57,10 +57,10 @@
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.RobertoRuizApp.NoActionBar">
-            <!--<intent-filter>
+            <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>-->
+            </intent-filter>
         </activity>
 
 
@@ -78,10 +78,6 @@
             android:name=".framework.view.activities.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER"/>
-            </intent-filter>
         </activity>
         <activity
             android:name=".framework.view.activities.EditProfileActivity"
