--- app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Eventos/Data.kt
@@ -1,5 +1,0 @@
-package com.example.kotlin.robertoruizapp.data.network.model.Eventos
-
-data class Data (
-    val documents: List<Document>
-)
