--- app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Eventos/Document.kt
@@ -1,12 +1,0 @@
-package com.example.kotlin.robertoruizapp.data.network.model.Eventos
-
-data class Document (
-
-    val ubicacion: String,
-    val _id: String,
-    val nombre: String,
-    val descripcion: String,
-    val fecha_Inicio: String,
-    val fecha_Fin: String,
-    val fotoUrl: String
-)
