--- app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Eventos/EventosObjeto.kt
@@ -1,7 +1,0 @@
-package com.example.kotlin.robertoruizapp.data.network.model.Eventos
-
-data class EventosObjeto (
-    val `data`: Data,
-    val results: Int?,
-    val status: String
-)
