+++ app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Events/Data.kt
@@ -1,0 +1,5 @@
+package com.example.kotlin.robertoruizapp.data.network.model.Events
+
+data class Data (
+    val documents: List<Document>
+)
