+++ app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Events/Document.kt
@@ -1,0 +1,12 @@
+package com.example.kotlin.robertoruizapp.data.network.model.Events
+
+data class Document (
+
+    val location: String,
+    val _id: String,
+    val name: String,
+    val description: String,
+    val startDate: String,
+    val endDate: String,
+    val ImageUrl: String
+)
