+++ app/src/main/java/com/example/kotlin/robertoruizapp/data/network/model/Events/EventObject.kt
@@ -1,0 +1,7 @@
+package com.example.kotlin.robertoruizapp.data.network.model.Events
+
+data class EventObject (
+    val `data`: Data,
+    val results: Int?,
+    val status: String
+)
