+++ app/src/main/java/com/example/kotlin/robertoruizapp/framework/view/activities/EventoClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.kotlin.robertoruizapp.framework.view.activities
 
-import com.example.kotlin.robertoruizapp.data.network.model.Eventos.Document
+import com.example.kotlin.robertoruizapp.data.network.model.Events.Document
 interface EventoClickListener {
     fun onClick(document: Document)
 }
