+++ app/src/main/java/com/example/kotlin/robertoruizapp/framework/view/fragments/FragmentoFeed.kt
@@ -1,10 +1,8 @@
 package com.example.kotlin.robertoruizapp.framework.view.fragments
 
-import android.content.Intent
 import android.os.Build
 import com.example.kotlin.robertoruizapp.databinding.FragmentoFeedBinding
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
